Extract a helper class
