+++ Tests/LumpikTests/ProcessSetTests.swift
@@ -11,22 +11,6 @@ import Foundation
 import Dispatch
 @testable import Lumpik
 
-// do not use with other thread
-final class SingleConnectionPool: ConnectablePool {
-    var redis: RedisStore
-    
-    init() {
-        redis = try! RedisStore(host: "localhost", port: 6379)
-    }
-
-    func borrow() throws -> RedisStore {
-        return redis
-    }
-    func checkin(_ connection: RedisStore) {
-        //
-    }
-}
-
 class ProcessSetTests: XCTestCase {
     static var allTests : [(String, (ProcessSetTests) -> () throws -> Void)] {
         return [
